pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "geobench"
version = "0.1.0"
authors = [
    { name="Adhitya Bhawiyuga", email="a.bhawiyuga@utwente.nl" },
    { name="Serkan Girgin", email="girgink@gmail.com"},
]
description = "A package and command-line utility to benchmark geospatial workflows."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
]
dependencies = [
    "jinja2",
    "pandas",
    "plotly",
    "psutil",
    "pyyaml",
    "requests",
]
keywords = [
    "geospatial analysis",
    "benchmarking",
    "performance monitoring",
]

[project.urls]
Homepage = "https://github.com/ITC-CRIB/geobench"
Source = "https://github.com/ITC-CRIB/geobench"
Issues = "https://github.com/ITC-CRIB/geobench/issues"
Changelog = "https://github.com/ITC-CRIB/geobench/commits"

[project.scripts]
geobench = "geobench.main:main"

[tool.hatch.build]
include = ["src/geobench/**/*"]

[tool.hatch.build.targets.wheel]
packages = ["src/geobench"]
